feat: Implement SignedDocumentCreator for PDF signing process
- Added SignedDocumentCreator class to manage the creation of signed PDF documents.
- Implemented methods for loading documents, verifying integrity, loading signatures, applying signatures, and sending signed documents via email.
- Introduced SecureCertificateManager for handling secure certificates and cryptographic signatures.
- Enhanced error handling and logging throughout the signing process.
- Added functionality to convert SVG signatures to images and incorporate them into the PDF.
- Included metadata and verification details in the generated certificates.
